App class as starting point
