+++ src/Bootstrap.php
@@ -37,15 +37,20 @@ class Bootstrap
 
     /**
      * Boots the framework.
-     * @throws \Exception|StopExecutionException
+     * @throws \Quantum\Exceptions\ControllerException
+     * @throws \Quantum\Exceptions\CsrfException
+     * @throws \Quantum\Exceptions\DiException
+     * @throws \Quantum\Exceptions\EnvException
+     * @throws \Quantum\Exceptions\LangException
+     * @throws \Quantum\Exceptions\LoaderException
+     * @throws \Quantum\Exceptions\MiddlewareException
+     * @throws \Quantum\Exceptions\ModuleLoaderException
+     * @throws \Quantum\Exceptions\RouteException
+     * @throws \ReflectionException
      */
     public static function run()
     {
         try {
-
-            self::loadCoreFuncations();
-
-            Di::loadDefinitions();
 
             $loader = Di::get(Loader::class);
 
@@ -71,26 +76,17 @@ class Bootstrap
             $loader->loadDir(base_dir() . DS . 'libraries');
 
             Lang::getInstance()
-                    ->setLang($request->getSegment(config()->get('lang_segment')))
-                    ->load($loader, $fs);
+                ->setLang($request->getSegment(config()->get('lang_segment')))
+                ->load($loader, $fs);
 
             MvcManager::runMvc($request, $response);
 
-            $response->send();
-            exit(0);
+            stop();
         } catch (StopExecutionException $e) {
             $response->send();
-        } catch (\Exception $e) {
-            echo $e->getMessage();
-            exit(1);
         }
+
     }
 
-    public static function loadCoreFuncations()
-    {
-        foreach (glob(HELPERS_DIR . DS . 'functions' . DS . '*.php') as $filename) {
-            require_once $filename;
-        }
-    }
 
 }
